service/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile ./service/requirements.in
#
asgiref==3.6.0
    # via django
attrs==23.1.0
    # via
    #   jsonschema
    #   referencing
build==0.10.0
    # via pip-tools
certifi==2022.12.7
    # via requests
charset-normalizer==3.0.1
    # via requests
click==8.1.3
    # via pip-tools
django==4.1.5
    # via
    #   -r ./service/requirements.in
    #   django-cors-headers
    #   djangorestframework
    #   drf-spectacular
django-cors-headers==3.13.0
    # via -r ./service/requirements.in
djangorestframework==3.14.0
    # via
    #   -r ./service/requirements.in
    #   drf-spectacular
drf-spectacular==0.26.4
    # via -r ./service/requirements.in
idna==3.4
    # via requests
inflection==0.5.1
    # via drf-spectacular
jsonschema==4.18.4
    # via drf-spectacular
jsonschema-specifications==2023.7.1
    # via jsonschema
packaging==23.0
    # via build
pip-tools==6.12.1
    # via -r ./service/requirements.in
psycopg2-binary==2.9.5
    # via -r ./service/requirements.in
pyproject-hooks==1.0.0
    # via build
pytz==2022.7.1
    # via djangorestframework
pyyaml==6.0.1
    # via drf-spectacular
referencing==0.30.0
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.28.2
    # via -r ./service/requirements.in
rpds-py==0.9.2
    # via
    #   jsonschema
    #   referencing
sqlparse==0.4.3
    # via django
tomli==2.0.1
    # via
    #   build
    #   pyproject-hooks
uritemplate==4.1.1
    # via drf-spectacular
urllib3==1.26.14
    # via requests
wheel==0.38.4
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
